REFACTOR test suite into `Feature` & `Unit` namespaces
BUMP CHANGELOG.md.
