+++ tests/Feature/LaravelConfigTest.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Sfneal\Helpers\Redis\Tests;
+namespace Sfneal\Helpers\Redis\Tests\Feature;
+
+use Sfneal\Helpers\Redis\Tests\TestCase;
 
 class LaravelConfigTest extends TestCase
 {
